--- a/ecc/bls12-377/fr/asm.go
+++ b/ecc/bls12-377/fr/asm.go
@@ -22,6 +22,8 @@ package fr
 import "golang.org/x/sys/cpu"
 
 var (
-	supportAdx = cpu.X86.HasADX && cpu.X86.HasBMI2
-	_          = supportAdx
+	supportAdx    = cpu.X86.HasADX && cpu.X86.HasBMI2
+	_             = supportAdx
+	supportAvx512 = cpu.X86.HasAVX512 && cpu.X86.HasAVX512DQ
+	_             = supportAvx512
 )