--- a/ecc/bls12-377/fr/element_ops_amd64.go
+++ b/ecc/bls12-377/fr/element_ops_amd64.go
@@ -83,7 +83,12 @@ func scalarMulVec(res, a, b *Element, n uint64)
 
 // Sum computes the sum of all elements in the vector.
 func (vector *Vector) Sum() (res Element) {
-	if len(*vector) == 0 {
+	n := uint64(len(*vector))
+	const minN = 16 * 7 // AVX512 slower than generic for small n
+	const maxN = (1 << 32) + 1
+	if !supportAvx512 || n <= minN || n >= maxN {
+		// call sumVecGeneric
+		sumVecGeneric(&res, *vector)
 		return
 	}
 	sumVec(&res, &(*vector)[0], uint64(len(*vector)))