--- a/ecc/bls12-377/fr/element_ops_amd64.go
+++ b/ecc/bls12-377/fr/element_ops_amd64.go
@@ -85,7 +85,7 @@ func scalarMulVec(res, a, b *Element, n uint64)
 func (vector *Vector) Sum() (res Element) {
 	n := uint64(len(*vector))
 	const minN = 16 * 7 // AVX512 slower than generic for small n
-	const maxN = (1 << 32) + 1
+	const maxN = (1 << 32) - 1
 	if !supportAvx512 || n <= minN || n >= maxN {
 		// call sumVecGeneric
 		sumVecGeneric(&res, *vector)