@gbotrel 2024-09-26
moving back to draft, will consolidate with other vecops (refactoring + better tests + cleaner)